requirements.txt
GitPython==3.1.8
Jinja2==2.11.2